Use multistage builds in a Dockerfile to compile and create an image from source
The multistage build lets us build the image inside of one container and
then copy over the results to a new image so the build dependencies
aren't needed in the final container.

The resulting container should be identical to the official built ones
for every release, but coming from source instead.